pyproject.toml
[project]
name = "search"
version = "3.0.19"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.12.15",
    "asyncpg==0.30.0",
    "confluent-kafka[protobuf,schemaregistry]==2.12.0",
    "elasticsearch[async]==8.19.1",
    "fastapi==0.119.0",
    "geojson==3.2.0",
    "grpcio==1.64.1",
    "numpy==2.3.4",
    "pandas==2.3.3",
    "protobuf==5.29.5",
    "pydantic==2.12.2",
    "pyjwt[crypto]==2.10.1",
    "python-dateutil==2.9.0.post0",
    "python-keycloak==5.8.1",
    "redis==6.4.0",
    "requests==2.32.5",
    "resistant-kafka-avataa==0.9.8b9",
    "sqlalchemy[asyncio]==2.0.44",
    "uvicorn==0.37.0",
    "xlsxwriter==3.2.9",
]

[dependency-groups]
dev = [
    "mypy==1.18.2",
    "pre-commit>=4.3.0",
]
lint = [
    "ruff==0.14.0",
]
tests = [
    "pytest==8.4.2",
    "httpx==0.28.1",
    "pytest-asyncio==1.2.0",
    "pytest-mock==3.15.1",
    "testcontainers==4.13.2",
    "asyncpg>=0.30.0",
]
security = [
    "pip-audit==2.9.0",
]
migrations = [
]
